requirements/development.txt
# Development dependencies - Updated June 2025
-r base.txt

# Development tools
django-debug-toolbar==5.2.0
django-extensions==4.1
ipython==8.31.0

# Code quality
black==25.1.0
isort==6.0.1
flake8==7.3.0
pylint==3.3.10
bandit==1.8.5
mypy==1.14.1
django-stubs==5.2.1

# Testing
pytest==8.4.1
pytest-django==4.11.1
pytest-cov==6.2.1
factory-boy==3.3.3

# Environment management
python-dotenv==1.0.1

# Pre-commit
pre-commit==4.0.1
